requirements.txt
click==6.7
decorator==4.2.1
dnspython==1.15.0
six==1.11.0
validators==0.12.1
